requirements.txt
# Anki相关
genanki==0.13.0

# 大模型API
openai>=2.6.0
anthropic==0.3.11

# 核心依赖
pydantic==2.5.0
PyYAML==6.0.1
click==8.1.7
colorama==0.4.6
tqdm==4.66.1
requests==2.31.0

# Web界面
Flask==2.3.3
Flask-Cors==4.0.0
werkzeug==2.3.7

# 音频处理（可选）
# gTTS==2.4.0
# edge-tts==6.1.3

# 开发和测试依赖（可选）
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.9.1
# flake8==6.1.0
# mypy==1.6.1